2nd and 3rd Encore bug fixed.
The nd and rd superscript was causing problems in the last scrape. The two
setlabels were interpreted as 2 and 3 with nd Encore and rd Encore appearing in
the setlist itself. This is now resolved.2nd and 3rd Encore bug fixed.

The nd and rd superscript was causing problems in the last scrape. The two
setlabels were interpreted as 2 and 3 with nd Encore and rd Encore appearing in
the setlist itself. This is now resolved.